implanted solid
